+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 CHANGELOG
 =========
 
+0.25.2
+------
+- Added `select` and `deselect` action for unconditinoally selecting or
+  deselecting a single item in `--multi` mode. Complements `toggle` action.
+- Built with Go 1.16
+
 0.25.1
 ------
 - Added `close` action